+++ src/gt4py/cartesian/backend/dace_backend.py
@@ -56,7 +56,9 @@ if TYPE_CHECKING:
 
 
 def _specialize_transient_strides(
-    sdfg: SDFG, layout_info: layout.LayoutInfo, replacement_dictionary: dict[str, str] | None = None
+    sdfg: SDFG,
+    layout_info: layout.LayoutInfo,
+    replacement_dictionary: dict[str, str] | None = None,
 ) -> None:
     # Find transients in this SDFG to specialize.
     stride_replacements = replace_strides(
@@ -385,14 +387,31 @@ class SDFGManager:
                 SDFGManager._loaded_sdfgs[path] = sdfg
                 return sdfg
 
-        # Create SDFG
+        # Get the schedule tree:
+        #  - we expect all 3-loops to be singled maps/for
+        #  - we expect the layout to be K-JI
+        #  - we expect all non-cartesian control flow to be innermost
         stree = self.schedule_tree()
 
-        if self.builder.backend.name.endswith("_kfirst"):
-            # re-order loops to match loops with memory layout
+        # Re-order cartesian loops to match loops with memory layout
+        # - layout is _always_ given I-J-K
+        # - layout is reversed, inner is higher numbers, e.g K-J-I is I=2, J=1, K=0
+        layout = self.builder.backend.storage_info["layout_map"](("I", "J", "K"))
+        if (layout[2] < layout[1] and layout[2] > layout[0]) or (
+            layout[2] > layout[1] and layout[2] < layout[0]
+        ):
+            raise NotImplementedError(
+                f"Layout (IJK:{layout}) is not implemented for dace:X backends"
+            )
+        if layout[0] < layout[1]:
+            flipper = passes.SwapHorizontalMaps()
+            flipper.visit(stree)
+
+        if layout[2] != 0:
             flipper = passes.PushVerticalMapDown()
             flipper.visit(stree)
 
+        # Create SDFG
         sdfg = stree.as_sdfg(
             validate=validate,
             simplify=simplify,
@@ -580,7 +599,10 @@ auto ${name}(const std::array<gt::uint_t, 3>& domain) {
                 config.Config.set("compiler", "cuda", "backend", value="hip")
             config.Config.set("compiler", "cuda", "max_concurrent_streams", value=-1)
             config.Config.set(
-                "compiler", "cuda", "default_block_size", value=gt_config.DACE_DEFAULT_BLOCK_SIZE
+                "compiler",
+                "cuda",
+                "default_block_size",
+                value=gt_config.DACE_DEFAULT_BLOCK_SIZE,
             )
             config.Config.set("compiler", "cpu", "openmp_sections", value=False)
             code_objects = sdfg.generate_code()
@@ -632,7 +654,11 @@ namespace gt = gridtools;
             for field_name, boundary in compute_k_boundary(stencil_ir).items()
         }
         offset_dict: dict[str, tuple[int, int, int]] = {
-            k: (max(-v[0][0], 0), max(-v[1][0], 0), k_origins[k] if k in k_origins else 0)
+            k: (
+                max(-v[0][0], 0),
+                max(-v[1][0], 0),
+                k_origins[k] if k in k_origins else 0,
+            )
             for k, v in field_extents.items()
         }
 
@@ -685,7 +711,9 @@ namespace gt = gridtools;
                 )
             )
             symbols[name] = fmt.format(
-                name=name, ndim=len(array.shape), origin=",".join(str(o) for o in origin)
+                name=name,
+                ndim=len(array.shape),
+                origin=",".join(str(o) for o in origin),
             )
 
         # the remaining arguments are variables and can be passed by name
@@ -821,7 +849,8 @@ class DaCePyExtModuleGenerator(PyExtModuleGenerator):
     def generate_class_members(self) -> str:
         res = super().generate_class_members()
         filepath = self.builder.module_path.joinpath(
-            os.path.dirname(self.builder.module_path), self.builder.module_name + ".sdfg"
+            os.path.dirname(self.builder.module_path),
+            self.builder.module_name + ".sdfg",
         )
         res += f'\nSDFG_PATH = "{filepath}"\n'
         return res
@@ -852,7 +881,7 @@ class DaceCPUBackend(BaseDaceBackend):
     storage_info: ClassVar[layout.LayoutInfo] = {
         "alignment": 1,
         "device": "cpu",
-        "layout_map": layout.layout_maker_factory((1, 2, 0)),
+        "layout_map": layout.layout_maker_factory((1, 2, 0)),  # Optimal loop order: K-I-J
         "is_optimal_layout": layout.layout_checker_factory(layout.layout_maker_factory((1, 2, 0))),
     }
     MODULE_GENERATOR_CLASS = DaCePyExtModuleGenerator
@@ -870,8 +899,26 @@ class DaceCPUKFirstBackend(BaseDaceBackend):
     storage_info: ClassVar[layout.LayoutInfo] = {
         "alignment": 1,
         "device": "cpu",
-        "layout_map": layout.layout_maker_factory((0, 1, 2)),
+        "layout_map": layout.layout_maker_factory((0, 1, 2)),  # Optimal loop order: I-J-K
         "is_optimal_layout": layout.layout_checker_factory(layout.layout_maker_factory((0, 1, 2))),
+    }
+    MODULE_GENERATOR_CLASS = DaCePyExtModuleGenerator
+
+    options = BaseGTBackend.GT_BACKEND_OPTS
+
+    def generate_extension(self) -> None:
+        return self.make_extension(uses_cuda=False)
+
+
+@register
+class DaceCPU_KJI(BaseDaceBackend):
+    name = "dace:cpu_KJI"
+    languages: ClassVar[dict] = {"computation": "c++", "bindings": ["python"]}
+    storage_info: ClassVar[layout.LayoutInfo] = {
+        "alignment": 1,
+        "device": "cpu",
+        "layout_map": layout.layout_maker_factory((2, 1, 0)),  # Optimal loop order: K-J-I
+        "is_optimal_layout": layout.layout_checker_factory(layout.layout_maker_factory((2, 1, 0))),
     }
     MODULE_GENERATOR_CLASS = DaCePyExtModuleGenerator
 
@@ -890,7 +937,7 @@ class DaceGPUBackend(BaseDaceBackend):
     storage_info: ClassVar[layout.LayoutInfo] = {
         "alignment": 32,
         "device": "gpu",
-        "layout_map": layout.layout_maker_factory((2, 1, 0)),
+        "layout_map": layout.layout_maker_factory((2, 1, 0)),  # Optimal loop order: K-J-I
         "is_optimal_layout": layout.layout_checker_factory(layout.layout_maker_factory((2, 1, 0))),
     }
     MODULE_GENERATOR_CLASS = DaCeCUDAPyExtModuleGenerator