+++ src/gt4py/next/common.py
@@ -32,6 +32,7 @@ from gt4py.eve.extended_typing import (
     Literal,
     NamedTuple,
     Never,
+    NoReturn,
     Optional,
     ParamSpec,
     Protocol,
@@ -49,6 +50,7 @@ from gt4py.eve.type_definitions import StrEnum
 
 
 DimT = TypeVar("DimT", bound="Dimension")  # , covariant=True)
+DimT_co = TypeVar("DimT_co", bound="Dimension", covariant=True)
 ShapeTs = TypeVarTuple("ShapeTs")
 
 
@@ -692,6 +694,11 @@ class Field(GTFieldInterface, Protocol[DimsT, core_defs.ScalarT]):
     def __str__(self) -> str:
         return f"⟨{self.domain!s} → {self.dtype}⟩"
 
+    def __bool__(self) -> NoReturn:
+        raise TypeError(
+            "The truth value of a Field is ambiguous. For one element Fields use '.as_scalar()'."
+        )
+
     @abc.abstractmethod
     def asnumpy(self) -> np.ndarray: ...
 
@@ -872,10 +879,10 @@ class ConnectivityKind(enum.Flag):
     - `ALTER_DIMS`: change the dimensions of the data field domain.
     - `ALTER_STRUCT`: transform structured of the data inside the field (non-compact transformation).
 
-    | Dims \ Struct |    No                    |    Yes                   |
-    | ------------- | ------------------------ | ------------------------ |
-    |   No          | Translation (I -> I)     | Reshuffling (I x K -> K) |
-    |   Yes         | Relocation (I -> I_half) | Remapping (V x V2E -> E) |
+    | Dims \\ Struct |    No                    |    Yes                   |
+    | -------------- | ------------------------ | ------------------------ |
+    |   No           | Translation (I -> I)     | Reshuffling (I x K -> K) |
+    |   Yes          | Relocation (I -> I_half) | Remapping (V x V2E -> E) |
 
     """
 
@@ -926,10 +933,10 @@ class NeighborConnectivityType(ConnectivityType):
 
 
 @runtime_checkable
-class Connectivity(Field[DimsT, core_defs.IntegralScalar], Protocol[DimsT, DimT]):
+class Connectivity(Field[DimsT, core_defs.IntegralScalar], Protocol[DimsT, DimT_co]):
     @property
     @abc.abstractmethod
-    def codomain(self) -> DimT:
+    def codomain(self) -> DimT_co:
         """
         The `codomain` is the set of all indices in a certain `Dimension`.