+++ tests/next_tests/unit_tests/test_containers.py
@@ -86,4 +86,39 @@ def test_make_container_constructor(
     constructed_container = constructor(nested_tuple)
 
     assert isinstance(constructed_container, container_type)
-    assert constructed_container == expected_container
+    for name in named_collections.elements_keys(expected_container):
+        constructed_value = getattr(constructed_container, name)
+        expected_value = getattr(expected_container, name)
+        if isinstance(expected_value, Field):
+            assert isinstance(constructed_value, Field)
+            assert constructed_value.domain == expected_value.domain
+            assert (constructed_value.asnumpy() == expected_value.asnumpy()).all()
+        else:
+            constructed_value == expected_value
+
+
+@pytest.mark.parametrize(
+    "first,second,expected",
+    [
+        (1, 41, 42),  # works without collections
+        ((1,), (41,), (42,)),  # works for simple tuples
+        (
+            # Note: we ignore typing and instantiate the collection with scalars for this test
+            cnc.DataclassNamedCollection(41, 43),
+            cnc.DataclassNamedCollection(1, -1),
+            cnc.DataclassNamedCollection(42, 42),
+        ),
+        (  # named collection in tuple
+            (cnc.SingleElementNamedTupleNamedCollection(x=1),),
+            (cnc.SingleElementNamedTupleNamedCollection(x=41),),
+            (cnc.SingleElementNamedTupleNamedCollection(x=42),),
+        ),
+    ],
+)
+def test_tree_map_named_collection_no_named_collection(first, second, expected):
+    @named_collections.tree_map_named_collection
+    def add(first, second):
+        return first + second
+
+    # Note: this comparison works only if we don't have fields in the collection as it relies on `==` comparison.
+    assert add(first, second) == expected